fix: OnyxUpdateManager cycle import warning
